ioctl requests are u_long not int, spotted via clang ok deraadt@ miod@

+++ sys/dev/pci/if_san_common.h
@@ -1,4 +1,4 @@
-/*	$OpenBSD: if_san_common.h,v 1.7 2005/04/01 21:42:36 canacar Exp $	*/
+/*	$OpenBSD: if_san_common.h,v 1.8 2005/11/08 20:23:42 canacar Exp $	*/
 
 /*-
  * Copyright (c) 2001-2004 Sangoma Technologies (SAN)
@@ -386,7 +386,7 @@ typedef struct sdla {
 	int(*iface_up) (struct ifnet*);
 	int(*iface_down) (struct ifnet*);
 	int(*iface_send) (struct mbuf* skb, struct ifnet*);
-	int(*iface_ioctl) (struct ifnet*, int, struct ifreq*);
+	int(*iface_ioctl) (struct ifnet*, u_long, struct ifreq*);
 
 	unsigned long	state_tick;	/* link state timestamp */
 	unsigned long	in_isr;		/* interrupt-in-service flag */
@@ -395,7 +395,7 @@ typedef struct sdla {
 	void(*isr)(struct sdla*);	/* interrupt service routine */
 	void(*poll)(struct sdla*);	/* polling routine */
 	int(*exec)(struct sdla*, void*, void*);
-	int(*ioctl) (struct ifnet*, int, struct ifreq*);
+	int(*ioctl) (struct ifnet*, u_long, struct ifreq*);
 
 	union {
 		sdla_xilinx_t	xilinx;
